[csharp] Add IList variant of Create{}Vector methods
This makes it convenient for applications to maintain reusable Lists for
vector offsets that you can .Clear() before appending offsets for variable
length vectors without requiring dynamic allocations once the list has
grown to a water-mark size.

C# Arrays can be awkward to reuse in many circumstances since their
capacity and element counts aren't decoupled as with a List. If you
are serializing types with variable numbers of elements you would likely
need lots of dynamic allocations to create suitably sized Arrays with
the previous API (or forgo the convenience API by unwrapping code
that works with a List).